aider: Changed the `proxy_pass` directive in the `/auth` location block to point to the correct authentication endpoint.
# Aider chat conversation:

USER: 192.168.65.1 - - [13/Nov/2023:21:39:52 +0000] "GET / HTTP/1.1" 500 177 "-" "curl/8.1.2" "-"
2023/11/13 21:39:52 [error] 28#28: *1 auth request unexpected status: 404 while sending to client, client: 192.168.65.1, server: test.localdev.me, request: "GET / HTTP/1.1", host: "test.localdev.me"
INFO:     192.168.112.2:49554 - "GET /auth HTTP/1.0" 404 Not Found

ASSISTANT: The error message indicates that the authentication request returned an unexpected status of 404. This means that the server could not find the requested resource. In this case, the resource is the authentication endpoint.

The issue could be due to the `proxy_pass` directive in the `/auth` location block in the `nginx/config/default.conf` file. The `proxy_pass` directive is used to define the address of the proxied server. This directive sets the address of the network server that will receive the request. The server needs to be running and listening on the address and port number specified in the `proxy_pass` directive.

Here's how to fix this:

```nginx
nginx/config/default.conf
<<<<<<< SEARCH
    location = /auth {
        internal;
        proxy_pass http://htaccess/authenticate;
        proxy_set_header Content-Length "";
        proxy_set_header X-Original-URI $request_uri;
        proxy_pass_request_body off;
    }
=======
    location = /auth {
        internal;
        proxy_pass http://htaccess/auth;
        proxy_set_header Content-Length "";
        proxy_set_header X-Original-URI $request_uri;
        proxy_pass_request_body off;
    }
>>>>>>> REPLACE
```

Please apply these changes and test your application again. If you're still getting a 404 error, let me know and we'll continue troubleshooting.